+++ xc3_model/src/material.rs
@@ -11,8 +11,7 @@ pub use xc3_lib::mxmd::{
 use crate::{
     ImageTexture, Sampler,
     shader_database::{
-        ChannelXyz, Operation, OutputExpr, Parameter, ParameterXyz, ProgramHash, ShaderDatabase,
-        ShaderProgram, Value,
+        ChannelXyz, Parameter, ParameterXyz, ProgramHash, ShaderDatabase, ShaderProgram,
     },
 };
 
@@ -606,25 +605,15 @@ fn get_shader_legacy<S: GetProgramHash>(
     // 2: normal (only xy)
     // 3: specular (alpha is spec?)
     // 4: depth (alpha is glossiness)
-    let mut exprs = program.exprs;
     let output_dependencies = if !is_single_output {
         program
             .output_dependencies
             .iter()
             .filter_map(|(k, v)| match k.as_str() {
                 // Ambient Occlusion
-                "o0.x" => {
-                    // Undo the multiply by 0.5 used for XCX and XCX DE.
-                    // This avoids needing to modify the actual database file.
-                    let const_index = exprs.len();
-                    exprs.push(OutputExpr::Value(Value::Float(2.0.into())));
-                    let index = exprs.len();
-                    exprs.push(OutputExpr::Func {
-                        op: Operation::Mul,
-                        args: vec![*v, const_index],
-                    });
-                    Some(("o2.z".into(), index))
-                }
+                // The actual in game shaders render RGB ambient lighting.
+                // This single channel ambient occlusion value is inferred.
+                "o2.z" => Some(("o2.z".into(), *v)),
                 // Color
                 "o1.x" => Some(("o0.x".into(), *v)),
                 "o1.y" => Some(("o0.y".into(), *v)),
@@ -673,7 +662,6 @@ fn get_shader_legacy<S: GetProgramHash>(
 
     Some(ShaderProgram {
         output_dependencies,
-        exprs,
         output_dependencies_xyz,
         ..program
     })