+++ CHANGELOG.md
@@ -8,11 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## unreleased
 ### Added
 * Added rendering and shader database support for model fragment shader alpha testing.
+* Added a `--game` arg to xc3_shader `shader-database` and `glsl-output-dependencies` to handle differences in shader outputs between games.
 
 ### Changed
 * Reworked xc3_model and xc3_shader database and material assignment types.
 * Changed shader assignments inferred from textures to merge XYZ channels for color and specular color.
 * Renamed the "tracing" feature for xc3_viewer and xc3_shader to "tracy".
+* Improved accuracy of material assignments for ambient occlusion for Xenoblade Chronicles X Definitive Edition.
 
 ### Fixed
 * Fixed an issue where map models would use the wrong alpha testing texture in some cases.